+++ opt/snobfit/test/test_snobfit.py
@@ -8,6 +8,7 @@ class TestSNOBFIT:
     def setup_class(cls):
         import SQSnobFit, logging, numpy
         SQSnobFit.log.setLevel(logging.DEBUG)
+        logging.getLogger('SKQ.SnobFit.minq').setLevel(logging.INFO)
 
     def reset(self):
         import SQSnobFit
@@ -112,7 +113,7 @@ class TestSNOBFIT:
 
             return -(c.dot(numpy.exp(-d)))
 
-        def run_Hartman6(self, initial, expected):
+        def run_Hartman6(self, initial, expected, options=None):
             self.reset()
 
             bounds = np.array([[0, 1], [0, 1], [0, 1], [0, 1], [0, 1], [0, 1]], dtype=float)
@@ -120,15 +121,21 @@ class TestSNOBFIT:
             x0 = np.array(initial)
 
             from SQSnobFit import optset
-            options = optset(maxmp=6+6)
+            if options is None:
+                options = optset(maxmp=6+6)
             result, history = SQSnobFit.minimize(Hartman6, x0, bounds, budget, options)
 
             assert np.round(sum(result.optpar)-sum(expected), 8) == 0
 
-      # note: results are still subtly different from the reference, but the most likely
-      # reason seems to be that the respective random number generators differ slightly
-        run_Hartman6(self, [],      (0.2077,  0.14892, 0.4829,  0.2725,  0.31493, 0.66138))
-        run_Hartman6(self, [0.5]*6, (0.20133, 0.1504,  0.47666, 0.27787, 0.3134,  0.65797))
+      # note: results are still subtly different from the reference, but the errors seem to
+      # accumulate over many iterations, not produced by a single step going wrong and the
+      # results here actually slightly outperform (?)
+        run_Hartman6(self, [],      (0.20687, 0.14968, 0.48076, 0.27357, 0.3145,  0.66129))
+        run_Hartman6(self, [0.5]*6, (0.21015, 0.12032, 0.46593, 0.27053, 0.30835, 0.66478))
+
+      # regression: the following used to fail
+        run_Hartman6(self, [],      (0.02423, 0.12107, 0.98254, 0.05482, 0.07433, 0.86491), {'maxmp' : 1})
+        run_Hartman6(self, [],      (0.24101, 0.16523, 0.43633, 0.28035, 0.31964, 0.64909), {'maxmp' : 2, 'maxfail' : 10})
 
     def test04_direct_call(self):
         """Direct call of a single iteration"""