+++ test/test_optimizers.py
@@ -5,20 +5,38 @@ import numpy as np
 sys.path = [os.pardir] + sys.path
 
 
+def f_easy_simple(x):
+    from math import sin
+
+    fv = np.inner(x, x)
+    fv *= 1 + 0.1*sin(10*(x[0]+x[1]))
+
+    return fv
+
+ref_results = {
+    f_easy_simple : {
+        'imfil'   : (-0.00681757, 0.00880743),
+        'snobfit' : (-0.00112, -0.00078),
+        'bobyqa'  : (-2.3883e-10, -1.3548e-10),
+    }
+}
+
+
 class TestOPTIMIZERS:
     def setup_class(cls):
         pass
 
+    def reset(self):
+        import SQSnobFit
+
+     # reset the random state for each method to get predictable results
+        SQSnobFit._gen_utils._randstate = np.random.RandomState(6)
+
+    def setup_method(self, method):
+        self.reset()
+
     def test01_availability(self):
         """Quick API check on available optimizers"""
-
-        def f_easy_simple(x):
-            from math import sin
-
-            fv = np.inner(x, x)
-            fv *= 1 + 0.1*sin(10*(x[0]+x[1]))
-
-            return fv
 
         import skquant.opt as skqopt
 
@@ -33,6 +51,7 @@ class TestOPTIMIZERS:
         result, history = \
              skqopt.minimize(f_easy_simple, x0, bounds, budget, method='imfil')
         assert type(result.optpar) == np.ndarray
+        assert np.round(sum(result.optpar)-sum(ref_results[f_easy_simple]['imfil']), 8) == 0.0
 
         # SnobFit
         from SQSnobFit import optset
@@ -40,7 +59,7 @@ class TestOPTIMIZERS:
         result, history = \
              skqopt.minimize(f_easy_simple, [], bounds, budget, method='snobfit', options=options)
         assert type(result.optpar) == np.ndarray
-        assert np.round(sum(result.optpar)-sum((-0.00112, -0.00078)), 8) == 0
+        assert np.round(sum(result.optpar)-sum(ref_results[f_easy_simple]['snobfit']), 8) == 0.0
 
         # Py-BOBYQA
         result, history = \
@@ -56,3 +75,77 @@ class TestOPTIMIZERS:
                  skqopt.minimize(f_easy_simple, x0, bounds, budget, method='orbit')
             assert type(result.optpar) == np.ndarray
             assert np.round(sum(result.optpar)-sum((0.00076624, 0.00060909)), 7) == 0
+
+    def test02_qiskit_interface(self):
+        """Verify usability of the Qiskit Aqua interface"""
+
+        try:
+            import qiskit.aqua
+        except ImportError:
+            py.test.skip("qiskit.aqua not available; not testing qiskit interface")
+
+        bounds = np.array([[-1, 1], [-1, 1]], dtype=float)
+        budget = 40
+        x0 = np.array([0.5, 0.5])
+
+      # ImFil
+        from skquant.interop.qiskit import ImFil
+
+        optimizer = ImFil(maxfun=budget)
+
+        ret = optimizer.optimize(num_vars=len(x0), objective_function=f_easy_simple, \
+                                 variable_bounds=bounds, initial_point=x0)
+        assert np.round(sum(ret[0])-sum(ref_results[f_easy_simple]['imfil']), 8) == 0.0
+
+      # SnobFit
+        from skquant.interop.qiskit import SnobFit
+
+        optimizer = SnobFit(maxfun=budget, maxmp=len(x0)+6)
+
+        ret = optimizer.optimize(num_vars=len(x0), objective_function=f_easy_simple, \
+                                 variable_bounds=bounds)
+        assert np.round(sum(ret[0])-sum(ref_results[f_easy_simple]['snobfit']), 8) == 0.0
+
+      # PyBobyqa
+        from skquant.interop.qiskit import PyBobyqa
+        optimizer = PyBobyqa(maxfun=budget)
+
+        ret = optimizer.optimize(num_vars=len(x0), objective_function=f_easy_simple, \
+                                 variable_bounds=bounds, initial_point=x0)
+        assert np.round(sum(ret[0])-sum(ref_results[f_easy_simple]['bobyqa']), 8) == 0.0
+
+    def test03_scipy_interface(self):
+        """Verify usability of the SciPy interface"""
+
+        try:
+            import scipy.optimize
+        except ImportError:
+            py.test.skip("scipy not available; not testing qiskit interface")
+
+        bounds = np.array([[-1, 1], [-1, 1]], dtype=float)
+        budget = 40
+        x0 = np.array([0.5, 0.5])
+
+      # ImFil
+        from skquant.interop.scipy import imfil
+
+        result = scipy.optimize.minimize(f_easy_simple, x0, method=imfil, bounds=bounds,
+                                         options={'budget' : budget})
+        assert type(result) == scipy.optimize.OptimizeResult
+        assert np.round(sum(result.x)-sum(ref_results[f_easy_simple]['imfil']), 8) == 0.0
+
+      # SnobFit
+        from skquant.interop.scipy import snobfit
+
+        result = scipy.optimize.minimize(f_easy_simple, [], method=snobfit, bounds=bounds,
+                                         options={'budget' : budget, 'maxmp' : len(x0)+6})
+        assert type(result) == scipy.optimize.OptimizeResult
+        assert np.round(sum(result.x)-sum(ref_results[f_easy_simple]['snobfit']), 8) == 0.0
+
+      # PyBobyqa
+        from skquant.interop.scipy import pybobyqa
+
+        result = scipy.optimize.minimize(f_easy_simple, x0, method=pybobyqa, bounds=bounds,
+                                         options={'budget' : budget})
+        assert type(result) == scipy.optimize.OptimizeResult
+        assert np.round(sum(result.x)-sum(ref_results[f_easy_simple]['bobyqa']), 8) == 0.0